Add files via upload
Input 2 CSV files outputs different correlation graphs of the entire dataset and the top 10% of the dataset. Need the files to have a similar column that you can compare unique entries with and need an additional column of the counts of each unique item in the column you are comparing. For example I have 2 csv files that has Barcode1 column in each of them and another column called Read_Counts in each of the files that is the number of reads I have for each Barcode1. I compare the abundance of Barcode1 in different steps of my cloning process to see if the abundance of unique things stays similar and if within each step there is skew.