Add files via upload
Treble Step1 Analysis locally in JN. designed to take inputs of .txt file of a list of designed DNA sequences and .fastq file of sequencing data. The notebook processes the reads looking for 3 barcodes and 1 designed DNA sequence based on common flanking regions and matching length requirements. Several Graphs and files are created to show quality distribution, abundance, and coverage. An initial filtering step of only keeping reads that have the correct flanking regions and have barcodes of all the correct lengths, and the designed DNA sequence being present in the design file is performed to create Map3. Map3 is then filtered to keep only reads that occur greater than or equal to a threshold number of read times that you can update. Then a filtering step is performed on the filtered Map3 to remove any reads that have BC3 (our reporter barcode) that map back to multiple HA (BC1-BC2 combinations) and to remove any other reads that have any of the 3 barcodes from that row.  (It is not important for our assay to remove the reverse aka HA that map to multiple BC3 because we collapse all those BC3 to one HA during the experimental analysis but you can modify the code to do so if you want). Finally Map4 is created which contains only rows with unique entry in the Cat column which is a combination of BC1-BC2-BC3-Designed DNA seq, it also adds an additional row to keep track of the number of reads for that unique Cat that appeared before collapsing. Cat and a few other barcode combinations (HA= BC1-BC2; HAR = BC1-BC2-BC3) are important for our assay, if you don't need them or need different barcode combinations, modify the code accordingly. A summary file is created at the end to summarize the reads removed at each step and check barcode diversity. 

The resulting Map4 is used as a Map1 look up table to link the three barcodes to the designed DNA sequence, because in later cloning steps the barcodes become more spread out where you cannot capture them on one Illumina seq read and you will only recover the barcode and not the full designed DNA seq.